latest sales put into chart
