+++ src/pages/chart.js
@@ -10,12 +10,31 @@ import {
 	Cell,
 	ResponsiveContainer,
 } from 'recharts';
-import { Grid, Typography } from '@material-ui/core';
-import { getSales, getTotalProfit, getTotalCustomer } from '../actions';
+import {
+	Grid,
+	Button,
+	Box,
+	Card,
+	CardHeader,
+	Chip,
+	Divider,
+	Table,
+	TableBody,
+	TableCell,
+	TableHead,
+	TableRow,
+} from '@material-ui/core';
+import {
+	getSales,
+	getTotalProfit,
+	getTotalCustomer,
+	getLatestSales,
+} from '../actions';
 import { useCarState } from '../hooks/use-car-state';
 import TotalProfit from '../components/total-profit';
 import TotalCustomers from '../components/total-customers';
 import Page from '../components/page';
+import { formatPrice } from '../utils/format-price';
 
 const Chart = () => {
 	const dispatch = useDispatch();
@@ -29,7 +48,8 @@ const Chart = () => {
 		dispatch,
 	]);
 	const [activeIndex, setActiveIndex] = useState(0);
-	const [activeItem, setActiveItem] = useState(sales[activeIndex]);
+	// eslint-disable-next-line no-unused-vars
+	const [_, setActiveItem] = useState(sales[activeIndex]);
 
 	const handleClick = (data, index) => setActiveIndex(index);
 
@@ -41,6 +61,14 @@ const Chart = () => {
 		setActiveItem(sales[activeIndex]);
 	}, [activeIndex, sales]);
 
+	const { latestSales } = useCarState();
+
+	const getLatestSalesCb = useCallback(() => dispatch(getLatestSales()), [
+		dispatch,
+	]);
+
+	useEffect(getLatestSalesCb, [getLatestSalesCb]);
+
 	return (
 		<Page title='Satış Grafiği'>
 			<Grid
@@ -50,51 +78,110 @@ const Chart = () => {
 				justify='center'
 				style={{ marginTop: '.5rem' }}
 			>
-				<Grid item container spacing={3} md={8} xs={12}>
-					<Grid item md={6} xs={12}>
+				<Grid item container spacing={3} md={10} xs={12}>
+					<Grid item md={4} xs={12}>
 						<TotalProfit />
 					</Grid>
-					<Grid item md={6} xs={12}>
+					<Grid item md={4} xs={12}>
+						<TotalCustomers />
+					</Grid>
+
+					{/* !TODO CHANGE THIS */}
+					<Grid item md={4} xs={12}>
 						<TotalCustomers />
 					</Grid>
 				</Grid>
-				<Grid item style={{ marginTop: '1rem' }}>
-					<Typography variant='h6' align='center'>
-						1 Aralık 2020 ve 1 Şubat 2021 Arası Satış Grafiği
-					</Typography>
-					<ResponsiveContainer height={400} maxWidth='100%'>
-						<LineChart
-							data={sales}
-							margin={{
-								top: 5,
-								right: 20,
-								bottom: 5,
-								left: 0,
-							}}
-						>
-							<CartesianGrid strokeDasharray='5 5' />
-							<XAxis dataKey='sale_date' />
-							<YAxis dateKey='count' />
-							<Tooltip />
-							<Line
-								type='monotone'
-								dataKey='count'
-								onClick={handleClick}
-							>
-								{sales.map((entry, index) => (
-									<Cell
-										cursor='pointer'
-										fill={
-											index === activeIndex
-												? '#82ca9d'
-												: '#8884d8'
-										}
-										key={`cell-${index}`}
-									/>
-								))}
-							</Line>
-						</LineChart>
-					</ResponsiveContainer>
+				<Grid
+					item
+					container
+					justify='center'
+					alignItems='center'
+					direction='row'
+					spacing={3}
+					style={{ marginTop: '1rem' }}
+				>
+					<Grid item md={5} sm={12}>
+						<Card>
+							<CardHeader
+								title='1 Aralık 2020 ve 1 Şubat 2021 Arası Satış
+								Grafiği'
+							/>
+							<Divider />
+							<ResponsiveContainer height={400} maxWidth='100%'>
+								<LineChart
+									data={sales}
+									style={{ left: '-1rem' }}
+								>
+									<CartesianGrid strokeDasharray='5 5' />
+									<XAxis dataKey='sale_date' />
+									<YAxis dateKey='count' />
+									<Tooltip />
+									<Line
+										type='monotone'
+										dataKey='count'
+										onClick={handleClick}
+									>
+										{sales.map((entry, index) => (
+											<Cell
+												cursor='pointer'
+												fill={
+													index === activeIndex
+														? '#82ca9d'
+														: '#8884d8'
+												}
+												key={`cell-${index}`}
+											/>
+										))}
+									</Line>
+								</LineChart>
+							</ResponsiveContainer>
+						</Card>
+					</Grid>
+					<Grid item md={5} sm={12}>
+						<Card>
+							<CardHeader title='Son Satışlar' />
+							<Divider />
+							<Box maxWidth='100%'>
+								<Table>
+									<TableHead>
+										<TableRow>
+											<TableCell>
+												Satış Numarası
+											</TableCell>
+											<TableCell>İlan Başlığı</TableCell>
+											<TableCell>Fiyat</TableCell>
+										</TableRow>
+									</TableHead>
+									<TableBody>
+										{latestSales.map((sale) => (
+											<TableRow hover key={sale.title}>
+												<TableCell>
+													{sale.serial_number}
+												</TableCell>
+												<TableCell>
+													{sale.title}
+												</TableCell>
+												<TableCell>
+													<Chip
+														color='primary'
+														label={formatPrice(
+															sale.purchase_price,
+														)}
+														size='small'
+													/>
+												</TableCell>
+											</TableRow>
+										))}
+									</TableBody>
+								</Table>
+							</Box>
+							<Box display='flex' justifyContent='flex-end' p={2}>
+								<Button variant='outlined' color='primary'>
+									Hepsini Gör
+								</Button>
+							</Box>
+						</Card>
+					</Grid>
 				</Grid>
 			</Grid>
 		</Page>
